extract API to submodule
Add docstrings to public methods
Remove unused methods
